refactor(api): split protocolStats tests by module
